+++ src/main/java/org/example/entity/invoice/UpdateInvoiceDTO.java
@@ -1,5 +1,6 @@
 package org.example.entity.invoice;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.UUID;
@@ -7,7 +8,7 @@ import java.util.UUID;
 public record UpdateInvoiceDTO(
     UUID invoiceId,
     LocalDateTime dueDate,
-    Float vatRate,
+    BigDecimal vatRate,
     List<InvoiceItemDTO> items,
     InvoiceStatus status
 ) {}
